fix(host-parasite): strict-typing annotations for ClaimGate adapters

+++ src/codontrace/genesis/host_parasite_campaign.py
@@ -238,7 +238,7 @@ def _stdev(values: Sequence[float]) -> float:
         return 0.0
     mean = sum(values) / len(values)
     var = sum((value - mean) ** 2 for value in values) / (len(values) - 1)
-    return round(var**0.5, 10)
+    return float(round(var**0.5, 10))
 
 
 def _ensure_overlap(host_tasks: Sequence[str], parasite_tasks: Sequence[str]) -> tuple[str, ...]:
